fixes
made it so after safari world is complete, ice and sea open up.
added all nodes in player scene to a group, so the portals can detect the group (it was detecting a child).
Made last portal in safari world teleport player back to home and set safari_finished in playerData to true